+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 ### Version 13.3
+- Fixed percentage calculation in formula analysis
 
 ### Version 13.2
 - Formula making api added "Complete Formula"