+++ releasenotes.md
@@ -1,3 +1,5 @@
 What's New in v13.3
-
+- Fixed percentage calculation errors in formula analysis.
+- Fixed popup link when clicking an ingredient in formula analysis.
+- Added visualization to highlight sub-materials exceeding usage limits.
 For a complete list of changes and previous versions, please refer to the CHANGELOG.md